+++ CHANGELOG.md
@@ -1,12 +1,45 @@
 # LzpCache Changelog
 
 ## Versão Atual
-1.1.0 - 06/06/2016
+1.2.0 - 08/06/2016
+
+
+## V 1.2.0
+-Documentação atualizada
+
+-Performance Otimizada
+
+-Modificado LzpCache -> Cache e Lozep -> Lzp
+
+-Modificado new Lozep\LzpCache -> new Lzp\Cache
+
+## V 1.1.1
+-Performance Otimizada
+
+-Melhora na Organização do código
+
+-Melhora na documentação
+
+-Novo parametro para DeleteAll($version)
+
+-Novo parametro para Create($config)
+
+-Modificação nas funções(veja mais nos exemplos):
+
+DeleteMultiples -> Retorna um array($nomecache=>$foiDeletado)
+
+CreateMultiples -> Retorna um array($nomecache=>$foiCriado)
+
+Delete -> Retorna true(sucesso), false(falha) ou null(não existe)
+
 
 ## V 1.1.0
 -Novo argumento opcional($cacheVersion)
+
 -Melhor documentação
+
 -Performance Otimizada
+
 -Nova função(cacheDirSize)
 
 ## V 1.0.0